Add GaussianPro

+++ data_extraction/data_extraction.py
@@ -209,7 +209,7 @@ def read_csvs(sources_file):
                         print('Downloaded', dataset)
                         data = StringIO(table)
                         df = pd.read_csv(data,dtype={
-                            'Size [Bytes]': 'Int64',
+                            'Size [Bytes]': 'float',
                         })
                         df['Submethod'] = df['Submethod'].apply(lambda x: '' if isinstance(x, str) and x.strip() == '' else x) # make empty submethod names empty
                         df['Submethod'] = df['Submethod'].apply(lambda x: ' ' + x if isinstance(x, str) and len(x)>0 and x != "Baseline" and x[0].isalpha() else x) #add space if starts with alphabetical character
